pyproject.toml
[project]
name = "realtime-english-transcriber"
version = "0.1.0"
description = "Real-time English speech-to-text transcription with Japanese translation"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "sounddevice>=0.5",
    "PyAudioWPatch>=0.2.12",
    "faster-whisper>=1.0",
    "torch>=2.0",
    "silero-vad>=5.1",
    "deepl>=1.16",
    "numpy>=1.24",
    "scipy>=1.11",
    "pycaw>=20230407",
    "comtypes>=1.2",
    "PyYAML>=6.0",
]

[project.scripts]
realtime-english = "src.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
